+++ R/makeGroupPlot.R
@@ -1,9 +1,12 @@
 makeGroupPlot <-
-function(fit, pl, pu, d = "best", lwd, xlab, ylab, fs = 12){
+function(fit, pl, pu, d = "best", lwd, xlab, ylab, fs = 12,
+         expertnames = NULL){
 	
   expert <- NULL # hack to avoid R CMD check NOTE
   
 	n.experts <- nrow(fit$vals)
+	
+	if(is.null(expertnames)){
 	
 	if(n.experts < 27){
 	  expertnames <- LETTERS[1:n.experts]
@@ -11,6 +14,8 @@ function(fit, pl, pu, d = "best", lwd, xlab, ylab, fs = 12){
 	
 	if(n.experts > 26){
 	  expertnames <- factor(1:n.experts)
+	}
+	
 	}
 	
 	x <- matrix(0, 200 * n.experts, 1)
@@ -23,7 +28,8 @@ function(fit, pl, pu, d = "best", lwd, xlab, ylab, fs = 12){
 		fx[(1+(i-1)*200):(i*200), 1] <-densitydata$fx
 	}
 	df1 <- data.frame(x = x, fx = fx, 
-	                  expert = rep(expertnames, each =200))
+	                  expert = factor(rep(expertnames, each =200),
+	                                  levels = expertnames))
 	p1 <- ggplot(df1, aes(x = x, y = fx, colour = expert))  +
 	  labs(x = xlab, y = ylab) +
 	  theme(text = element_text(size = fs))