+++ R/plotTertiles.R
@@ -28,14 +28,20 @@
 #' }
 #' @export
 
-plotTertiles <- function(vals, lower, upper, fs = 12, percentages = FALSE){
+plotTertiles <- function(vals, lower, upper, fs = 12,
+                         percentages = FALSE,
+                         expertnames = NULL){
   
   low <- L <- T1 <- M <- T2 <- U <- enumber <- NULL # hack to pass CRAN check
   
   
   n.experts <- ncol(vals)
-  expert <-factor(LETTERS[1 : n.experts], levels = LETTERS[n.experts : 1])
   
+  if(is.null(expertnames)){
+  expert <-factor(LETTERS[1 : n.experts], levels = LETTERS[n.experts : 1])
+  }else{
+    expert <- factor(expertnames, levels = expertnames)
+  }
   cols <- gg_color_hue(3)
   
   df1 <- data.frame(cbind(lower, t(vals), upper))