+++ NEWS.md
@@ -1,6 +1,10 @@
 SHELF v1.7.0.9000 (2020-05-26) 
 ==============================
 
+* plotfit() will now use the expert names, if provided in fitdist()
+
+* elicitMultiple(): can now click on and edit expert names.
+
 * new function rlinearpool() for sampling from a weighted linear pool
 
 * bug fixed in copulaSample. Was rounding samples to 3 s.f. Increased to 8. Mistake in help file corrected, regarding syntax for distribution names.